Adding action information in 7.1 (Part 4)
Fixed an issue where Scholar's LB icon and action did not match.
Fixed an issue where the barrier deployed with Deploy Tactics was not calculated (calculated as if there was no secret).
Fixed an issue where the barrier was not calculated with Red Mage's melee combo.
Added Red Sands Rest Water as a HoT.
Added types of effect flags.